Fix qemu compilation on Xcode 12.4
Recent version of MacOS 11.3 combined with XCode 12.4 fail to compile
qemu. The VERSION file in the root of qemu repository gets included as a
header instead of the header in system includes. The reason for this is
not clear yet. But for the time being, this patch will help those who
can't install qemu from this Formula.